Minor bug fix

+++ app/src/main/java/binar/and3/kelompok1/secondhand/data/api/seller/GetProductByIdResponse.kt
@@ -6,7 +6,7 @@ data class GetProductByIdResponse(
     @SerializedName("id") var id: Int? = null,
     @SerializedName("name") var name: String? = null,
     @SerializedName("description") var description: String? = null,
-    @SerializedName("base_price") var basePrice: Long? = null,
+    @SerializedName("base_price") var basePrice: Int? = null,
     @SerializedName("image_url") var imageUrl: String? = null,
     @SerializedName("image_name") var imageName: String? = null,
     @SerializedName("location") var location: String? = null,
